Moved importer into new files
