+++ beets/importer/tasks.py
@@ -12,44 +12,31 @@
 # The above copyright notice and this permission notice shall be
 # included in all copies or substantial portions of the Software.
 
-"""Provides the basic, interface-agnostic workflow for importing and
-autotagging music files.
-"""
-
 from __future__ import annotations
 
-import itertools
+import logging
 import os
-import pickle
 import re
 import shutil
 import time
-from bisect import bisect_left, insort
 from collections import defaultdict
-from dataclasses import dataclass
 from enum import Enum
 from tempfile import mkdtemp
-from typing import Callable, Iterable, Sequence
+from typing import TYPE_CHECKING, Callable, Iterable, Sequence
 
 import mediafile
 
-from beets import autotag, config, dbcore, library, logging, plugins, util
-from beets.util import (
-    MoveOperation,
-    ancestry,
-    displayable_path,
-    normpath,
-    pipeline,
-    sorted_walk,
-    syspath,
-)
+from beets import autotag, config, dbcore, library, plugins, util
 
-action = Enum("action", ["SKIP", "ASIS", "TRACKS", "APPLY", "ALBUMS", "RETAG"])
-# The RETAG action represents "don't apply any match, but do record
-# new metadata". It's not reachable via the standard command prompt but
-# can be used by plugins.
+from .state import ImportState
 
-QUEUE_SIZE = 128
+if TYPE_CHECKING:
+    from .session import ImportSession
+
+# Global logger.
+log = logging.getLogger("beets")
+
+
 SINGLE_ARTIST_THRESH = 0.25
 
 # Usually flexible attributes are preserved (i.e., not updated) during
@@ -74,9 +61,11 @@ REIMPORT_FRESH_FIELDS_ITEM = list(REIMPORT_FRESH_FIELDS_ALBUM)
 # Global logger.
 log = logging.getLogger("beets")
 
-# Here for now to allow for a easy replace later on
-# once we can move to a PathLike
-PathBytes = bytes
+
+action = Enum("action", ["SKIP", "ASIS", "TRACKS", "APPLY", "ALBUMS", "RETAG"])
+# The RETAG action represents "don't apply any match, but do record
+# new metadata". It's not reachable via the standard command prompt but
+# can be used by plugins.
 
 
 class ImportAbortError(Exception):
@@ -85,395 +74,20 @@ class ImportAbortError(Exception):
     pass
 
 
-@dataclass
-class ImportState:
-    """Representing the progress of an import task.
-
-    Opens the state file on creation of the class. If you want
-    to ensure the state is written to disk, you should use the
-    context manager protocol.
-
-    Tagprogress allows long tagging tasks to be resumed when they pause.
-
-    Taghistory is a utility for manipulating the "incremental" import log.
-    This keeps track of all directories that were ever imported, which
-    allows the importer to only import new stuff.
-
-    Usage
-    -----
-    ```
-    # Readonly
-    progress = ImportState().tagprogress
-
-    # Read and write
-    with ImportState() as state:
-        state["key"] = "value"
-    ```
-    """
-
-    tagprogress: dict[PathBytes, list[PathBytes]]
-    taghistory: set[tuple[PathBytes, ...]]
-    path: PathBytes
-
-    def __init__(self, readonly=False, path: PathBytes | None = None):
-        self.path = path or os.fsencode(config["statefile"].as_filename())
-        self.tagprogress = {}
-        self.taghistory = set()
-        self._open()
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self._save()
-
-    def _open(
-        self,
-    ):
-        try:
-            with open(self.path, "rb") as f:
-                state = pickle.load(f)
-                # Read the states
-                self.tagprogress = state.get("tagprogress", {})
-                self.taghistory = state.get("taghistory", set())
-        except Exception as exc:
-            # The `pickle` module can emit all sorts of exceptions during
-            # unpickling, including ImportError. We use a catch-all
-            # exception to avoid enumerating them all (the docs don't even have a
-            # full list!).
-            log.debug("state file could not be read: {0}", exc)
-
-    def _save(self):
-        try:
-            with open(self.path, "wb") as f:
-                pickle.dump(
-                    {
-                        "tagprogress": self.tagprogress,
-                        "taghistory": self.taghistory,
-                    },
-                    f,
-                )
-        except OSError as exc:
-            log.error("state file could not be written: {0}", exc)
-
-    # -------------------------------- Tagprogress ------------------------------- #
-
-    def progress_add(self, toppath: PathBytes, *paths: PathBytes):
-        """Record that the files under all of the `paths` have been imported
-        under `toppath`.
-        """
-        with self as state:
-            imported = state.tagprogress.setdefault(toppath, [])
-            for path in paths:
-                if imported and imported[-1] <= path:
-                    imported.append(path)
-                else:
-                    insort(imported, path)
-
-    def progress_has_element(self, toppath: PathBytes, path: PathBytes) -> bool:
-        """Return whether `path` has been imported in `toppath`."""
-        imported = self.tagprogress.get(toppath, [])
-        i = bisect_left(imported, path)
-        return i != len(imported) and imported[i] == path
-
-    def progress_has(self, toppath: PathBytes) -> bool:
-        """Return `True` if there exist paths that have already been
-        imported under `toppath`.
-        """
-        return toppath in self.tagprogress
-
-    def progress_reset(self, toppath: PathBytes | None):
-        """Reset the progress for `toppath`."""
-        with self as state:
-            if toppath in state.tagprogress:
-                del state.tagprogress[toppath]
-
-    # -------------------------------- Taghistory -------------------------------- #
-
-    def history_add(self, paths: list[PathBytes]):
-        """Add the paths to the history."""
-        with self as state:
-            state.taghistory.add(tuple(paths))
-
-
-class ImportSession:
-    """Controls an import action. Subclasses should implement methods to
-    communicate with the user or otherwise make decisions.
-    """
-
-    logger: logging.Logger
-    paths: list[PathBytes]
-    lib: library.Library
-
-    _is_resuming: dict[bytes, bool]
-    _merged_items: set[PathBytes]
-    _merged_dirs: set[PathBytes]
-
-    def __init__(
-        self,
-        lib: library.Library,
-        loghandler: logging.Handler | None,
-        paths: Sequence[PathBytes] | None,
-        query: dbcore.Query | None,
-    ):
-        """Create a session.
-
-        Parameters
-        ----------
-        lib : library.Library
-            The library instance to which items will be imported.
-        loghandler : logging.Handler or None
-            A logging handler to use for the session's logger. If None, a
-            NullHandler will be used.
-        paths : os.PathLike or None
-            The paths to be imported.
-        query : dbcore.Query or None
-            A query to filter items for import.
-        """
-        self.lib = lib
-        self.logger = self._setup_logging(loghandler)
-        self.query = query
-        self._is_resuming = {}
-        self._merged_items = set()
-        self._merged_dirs = set()
-
-        # Normalize the paths.
-        self.paths = list(map(normpath, paths or []))
-
-    def _setup_logging(self, loghandler: logging.Handler | None):
-        logger = logging.getLogger(__name__)
-        logger.propagate = False
-        if not loghandler:
-            loghandler = logging.NullHandler()
-        logger.handlers = [loghandler]
-        return logger
-
-    def set_config(self, config):
-        """Set `config` property from global import config and make
-        implied changes.
-        """
-        # FIXME: Maybe this function should not exist and should instead
-        # provide "decision wrappers" like "should_resume()", etc.
-        iconfig = dict(config)
-        self.config = iconfig
-
-        # Incremental and progress are mutually exclusive.
-        if iconfig["incremental"]:
-            iconfig["resume"] = False
-
-        # When based on a query instead of directories, never
-        # save progress or try to resume.
-        if self.query is not None:
-            iconfig["resume"] = False
-            iconfig["incremental"] = False
-
-        if iconfig["reflink"]:
-            iconfig["reflink"] = iconfig["reflink"].as_choice(
-                ["auto", True, False]
-            )
-
-        # Copy, move, reflink, link, and hardlink are mutually exclusive.
-        if iconfig["move"]:
-            iconfig["copy"] = False
-            iconfig["link"] = False
-            iconfig["hardlink"] = False
-            iconfig["reflink"] = False
-        elif iconfig["link"]:
-            iconfig["copy"] = False
-            iconfig["move"] = False
-            iconfig["hardlink"] = False
-            iconfig["reflink"] = False
-        elif iconfig["hardlink"]:
-            iconfig["copy"] = False
-            iconfig["move"] = False
-            iconfig["link"] = False
-            iconfig["reflink"] = False
-        elif iconfig["reflink"]:
-            iconfig["copy"] = False
-            iconfig["move"] = False
-            iconfig["link"] = False
-            iconfig["hardlink"] = False
-
-        # Only delete when copying.
-        if not iconfig["copy"]:
-            iconfig["delete"] = False
-
-        self.want_resume = config["resume"].as_choice([True, False, "ask"])
-
-    def tag_log(self, status, paths: Sequence[PathBytes]):
-        """Log a message about a given album to the importer log. The status
-        should reflect the reason the album couldn't be tagged.
-        """
-        self.logger.info("{0} {1}", status, displayable_path(paths))
-
-    def log_choice(self, task: ImportTask, duplicate=False):
-        """Logs the task's current choice if it should be logged. If
-        ``duplicate``, then this is a secondary choice after a duplicate was
-        detected and a decision was made.
-        """
-        paths = task.paths
-        if duplicate:
-            # Duplicate: log all three choices (skip, keep both, and trump).
-            if task.should_remove_duplicates:
-                self.tag_log("duplicate-replace", paths)
-            elif task.choice_flag in (action.ASIS, action.APPLY):
-                self.tag_log("duplicate-keep", paths)
-            elif task.choice_flag is (action.SKIP):
-                self.tag_log("duplicate-skip", paths)
-        else:
-            # Non-duplicate: log "skip" and "asis" choices.
-            if task.choice_flag is action.ASIS:
-                self.tag_log("asis", paths)
-            elif task.choice_flag is action.SKIP:
-                self.tag_log("skip", paths)
-
-    def should_resume(self, path: PathBytes):
-        raise NotImplementedError
-
-    def choose_match(self, task: ImportTask):
-        raise NotImplementedError
-
-    def resolve_duplicate(self, task: ImportTask, found_duplicates):
-        raise NotImplementedError
-
-    def choose_item(self, task: ImportTask):
-        raise NotImplementedError
-
-    def run(self):
-        """Run the import task."""
-        self.logger.info("import started {0}", time.asctime())
-        self.set_config(config["import"])
-
-        # Set up the pipeline.
-        if self.query is None:
-            stages = [read_tasks(self)]
-        else:
-            stages = [query_tasks(self)]
-
-        # In pretend mode, just log what would otherwise be imported.
-        if self.config["pretend"]:
-            stages += [log_files(self)]
-        else:
-            if self.config["group_albums"] and not self.config["singletons"]:
-                # Split directory tasks into one task for each album.
-                stages += [group_albums(self)]
-
-            # These stages either talk to the user to get a decision or,
-            # in the case of a non-autotagged import, just choose to
-            # import everything as-is. In *both* cases, these stages
-            # also add the music to the library database, so later
-            # stages need to read and write data from there.
-            if self.config["autotag"]:
-                stages += [lookup_candidates(self), user_query(self)]
-            else:
-                stages += [import_asis(self)]
-
-            # Plugin stages.
-            for stage_func in plugins.early_import_stages():
-                stages.append(plugin_stage(self, stage_func))
-            for stage_func in plugins.import_stages():
-                stages.append(plugin_stage(self, stage_func))
-
-            stages += [manipulate_files(self)]
-
-        pl = pipeline.Pipeline(stages)
-
-        # Run the pipeline.
-        plugins.send("import_begin", session=self)
-        try:
-            if config["threaded"]:
-                pl.run_parallel(QUEUE_SIZE)
-            else:
-                pl.run_sequential()
-        except ImportAbortError:
-            # User aborted operation. Silently stop.
-            pass
-
-    # Incremental and resumed imports
-
-    def already_imported(self, toppath: PathBytes, paths: Sequence[PathBytes]):
-        """Returns true if the files belonging to this task have already
-        been imported in a previous session.
-        """
-        if self.is_resuming(toppath) and all(
-            [ImportState().progress_has_element(toppath, p) for p in paths]
-        ):
-            return True
-        if self.config["incremental"] and tuple(paths) in self.history_dirs:
-            return True
-
-        return False
-
-    _history_dirs = None
-
-    @property
-    def history_dirs(self) -> set[tuple[PathBytes, ...]]:
-        # FIXME: This could be simplified to a cached property
-        if self._history_dirs is None:
-            self._history_dirs = ImportState().taghistory
-        return self._history_dirs
-
-    def already_merged(self, paths: Sequence[PathBytes]):
-        """Returns true if all the paths being imported were part of a merge
-        during previous tasks.
-        """
-        for path in paths:
-            if path not in self._merged_items and path not in self._merged_dirs:
-                return False
-        return True
-
-    def mark_merged(self, paths: Sequence[PathBytes]):
-        """Mark paths and directories as merged for future reimport tasks."""
-        self._merged_items.update(paths)
-        dirs = {
-            os.path.dirname(path) if os.path.isfile(syspath(path)) else path
-            for path in paths
-        }
-        self._merged_dirs.update(dirs)
-
-    def is_resuming(self, toppath: PathBytes):
-        """Return `True` if user wants to resume import of this path.
-
-        You have to call `ask_resume` first to determine the return value.
-        """
-        return self._is_resuming.get(toppath, False)
-
-    def ask_resume(self, toppath: PathBytes):
-        """If import of `toppath` was aborted in an earlier session, ask
-        user if they want to resume the import.
-
-        Determines the return value of `is_resuming(toppath)`.
-        """
-        if self.want_resume and ImportState().progress_has(toppath):
-            # Either accept immediately or prompt for input to decide.
-            if self.want_resume is True or self.should_resume(toppath):
-                log.warning(
-                    "Resuming interrupted import of {0}",
-                    util.displayable_path(toppath),
-                )
-                self._is_resuming[toppath] = True
-            else:
-                # Clear progress; we're starting from the top.
-                ImportState().progress_reset(toppath)
-
-
-# The importer task class.
-
-
 class BaseImportTask:
     """An abstract base class for importer tasks.
 
     Tasks flow through the importer pipeline. Each stage can update
     them."""
 
-    toppath: PathBytes | None
-    paths: list[PathBytes]
+    toppath: util.PathBytes | None
+    paths: list[util.PathBytes]
     items: list[library.Item]
 
     def __init__(
         self,
-        toppath: PathBytes | None,
-        paths: Iterable[PathBytes] | None,
+        toppath: util.PathBytes | None,
+        paths: Iterable[util.PathBytes] | None,
         items: Iterable[library.Item] | None,
     ):
         """Create a task. The primary fields that define a task are:
@@ -539,8 +153,8 @@ class ImportTask(BaseImportTask):
 
     def __init__(
         self,
-        toppath: PathBytes | None,
-        paths: Iterable[PathBytes] | None,
+        toppath: util.PathBytes | None,
+        paths: Iterable[util.PathBytes] | None,
         items: Iterable[library.Item] | None,
     ):
         super().__init__(toppath, paths, items)
@@ -662,7 +276,7 @@ class ImportTask(BaseImportTask):
             value = str(view.get())
             log.debug(
                 "Set field {1}={2} for {0}",
-                displayable_path(self.paths),
+                util.displayable_path(self.paths),
                 field,
                 value,
             )
@@ -823,7 +437,7 @@ class ImportTask(BaseImportTask):
     def manipulate_files(
         self,
         session: ImportSession,
-        operation: MoveOperation | None = None,
+        operation: util.MoveOperation | None = None,
         write=False,
     ):
         """Copy, move, link, hardlink or reflink (depending on `operation`)
@@ -838,7 +452,7 @@ class ImportTask(BaseImportTask):
         items = self.imported_items()
         # Save the original paths of all items for deletion and pruning
         # in the next step (finalization).
-        self.old_paths: list[PathBytes] = [item.path for item in items]
+        self.old_paths: list[util.PathBytes] = [item.path for item in items]
         for item in items:
             if operation is not None:
                 # In copy and link modes, treat re-imports specially:
@@ -846,7 +460,7 @@ class ImportTask(BaseImportTask):
                 # copied/moved as usual).
                 old_path = item.path
                 if (
-                    operation != MoveOperation.MOVE
+                    operation != util.MoveOperation.MOVE
                     and self.replaced_items[item]
                     and session.lib.directory in util.ancestry(old_path)
                 ):
@@ -893,11 +507,13 @@ class ImportTask(BaseImportTask):
         and `replaced_albums` dictionaries.
         """
         self.replaced_items = defaultdict(list)
-        self.replaced_albums: dict[PathBytes, library.Album] = defaultdict()
+        self.replaced_albums: dict[util.PathBytes, library.Album] = (
+            defaultdict()
+        )
         replaced_album_ids = set()
         for item in self.imported_items():
             dup_items = list(
-                lib.items(dbcore.query.BytesQuery("path", item.path))
+                lib.items(query=dbcore.query.BytesQuery("path", item.path))
             )
             self.replaced_items[item] = dup_items
             for dup_item in dup_items:
@@ -938,7 +554,7 @@ class ImportTask(BaseImportTask):
                     noun,
                     new_obj.id,
                     overwritten_fields,
-                    displayable_path(new_obj.path),
+                    util.displayable_path(new_obj.path),
                 )
                 for key in overwritten_fields:
                     del existing_fields[key]
@@ -960,14 +576,14 @@ class ImportTask(BaseImportTask):
                     "Reimported album {}. Preserving attribute ['added']. "
                     "Path: {}",
                     self.album.id,
-                    displayable_path(self.album.path),
+                    util.displayable_path(self.album.path),
                 )
                 log.debug(
                     "Reimported album {}. Preserving flexible attributes {}. "
                     "Path: {}",
                     self.album.id,
                     list(album_fields.keys()),
-                    displayable_path(self.album.path),
+                    util.displayable_path(self.album.path),
                 )
 
         for item in self.imported_items():
@@ -979,7 +595,7 @@ class ImportTask(BaseImportTask):
                         "Reimported item {}. Preserving attribute ['added']. "
                         "Path: {}",
                         item.id,
-                        displayable_path(item.path),
+                        util.displayable_path(item.path),
                     )
                 item_fields = _reduce_and_log(
                     item, dup_item._values_flex, REIMPORT_FRESH_FIELDS_ITEM
@@ -990,7 +606,7 @@ class ImportTask(BaseImportTask):
                     "Path: {}",
                     item.id,
                     list(item_fields.keys()),
-                    displayable_path(item.path),
+                    util.displayable_path(item.path),
                 )
                 item.store()
 
@@ -1003,7 +619,7 @@ class ImportTask(BaseImportTask):
                 log.debug(
                     "Replacing item {0}: {1}",
                     dup_item.id,
-                    displayable_path(item.path),
+                    util.displayable_path(item.path),
                 )
                 dup_item.remove()
         log.debug(
@@ -1033,7 +649,7 @@ class ImportTask(BaseImportTask):
         the file still exists, no pruning is performed, so it's safe to
         call when the file in question may not have been removed.
         """
-        if self.toppath and not os.path.exists(syspath(filename)):
+        if self.toppath and not os.path.exists(util.syspath(filename)):
             util.prune_dirs(
                 os.path.dirname(filename),
                 self.toppath,
@@ -1044,7 +660,7 @@ class ImportTask(BaseImportTask):
 class SingletonImportTask(ImportTask):
     """ImportTask for a single track that is not associated to an album."""
 
-    def __init__(self, toppath: PathBytes | None, item: library.Item):
+    def __init__(self, toppath: util.PathBytes | None, item: library.Item):
         super().__init__(toppath, [item.path], [item])
         self.item = item
         self.is_album = False
@@ -1127,7 +743,7 @@ class SingletonImportTask(ImportTask):
             value = str(view.get())
             log.debug(
                 "Set field {1}={2} for {0}",
-                displayable_path(self.paths),
+                util.displayable_path(self.paths),
                 field,
                 value,
             )
@@ -1246,9 +862,9 @@ class ArchiveImportTask(SentinelImportTask):
         if self.extracted and self.toppath:
             log.debug(
                 "Removing extracted directory: {0}",
-                displayable_path(self.toppath),
+                util.displayable_path(self.toppath),
             )
-            shutil.rmtree(syspath(self.toppath))
+            shutil.rmtree(util.syspath(self.toppath))
 
     def extract(self):
         """Extracts the archive to a temporary directory and sets
@@ -1289,7 +905,7 @@ class ImportTaskFactory:
     indicated by a path.
     """
 
-    def __init__(self, toppath: PathBytes, session: ImportSession):
+    def __init__(self, toppath: util.PathBytes, session: ImportSession):
         """Create a new task factory.
 
         `toppath` is the user-specified path to search for music to
@@ -1300,7 +916,7 @@ class ImportTaskFactory:
         self.session = session
         self.skipped = 0  # Skipped due to incremental/resume.
         self.imported = 0  # "Real" tasks created.
-        self.is_archive = ArchiveImportTask.is_archive(syspath(toppath))
+        self.is_archive = ArchiveImportTask.is_archive(util.syspath(toppath))
 
     def tasks(self):
         """Yield all import tasks for music found in the user-specified
@@ -1362,7 +978,7 @@ class ImportTaskFactory:
         single track when `toppath` is a file, a single directory in
         `flat` mode.
         """
-        if not os.path.isdir(syspath(self.toppath)):
+        if not os.path.isdir(util.syspath(self.toppath)):
             yield [self.toppath], [self.toppath]
         elif self.session.config["flat"]:
             paths = []
@@ -1373,11 +989,12 @@ class ImportTaskFactory:
             for dirs, paths in albums_in_dir(self.toppath):
                 yield dirs, paths
 
-    def singleton(self, path: PathBytes):
+    def singleton(self, path: util.PathBytes):
         """Return a `SingletonImportTask` for the music file."""
         if self.session.already_imported(self.toppath, [path]):
             log.debug(
-                "Skipping previously-imported path: {0}", displayable_path(path)
+                "Skipping previously-imported path: {0}",
+                util.displayable_path(path),
             )
             self.skipped += 1
             return None
@@ -1388,7 +1005,7 @@ class ImportTaskFactory:
         else:
             return None
 
-    def album(self, paths: Iterable[PathBytes], dirs=None):
+    def album(self, paths: Iterable[util.PathBytes], dirs=None):
         """Return a `ImportTask` with all media files from paths.
 
         `dirs` is a list of parent directories used to record already
@@ -1400,7 +1017,8 @@ class ImportTaskFactory:
 
         if self.session.already_imported(self.toppath, dirs):
             log.debug(
-                "Skipping previously-imported path: {0}", displayable_path(dirs)
+                "Skipping previously-imported path: {0}",
+                util.displayable_path(dirs),
             )
             self.skipped += 1
             return None
@@ -1414,7 +1032,7 @@ class ImportTaskFactory:
         else:
             return None
 
-    def sentinel(self, paths: Iterable[PathBytes] | None = None):
+    def sentinel(self, paths: Iterable[util.PathBytes] | None = None):
         """Return a `SentinelImportTask` indicating the end of a
         top-level directory import.
         """
@@ -1436,7 +1054,9 @@ class ImportTaskFactory:
             )
             return
 
-        log.debug("Extracting archive: {0}", displayable_path(self.toppath))
+        log.debug(
+            "Extracting archive: {0}", util.displayable_path(self.toppath)
+        )
         archive_task = ArchiveImportTask(self.toppath)
         try:
             archive_task.extract()
@@ -1449,7 +1069,7 @@ class ImportTaskFactory:
         log.debug("Archive extracted to: {0}", self.toppath)
         return archive_task
 
-    def read_item(self, path: PathBytes):
+    def read_item(self, path: util.PathBytes):
         """Return an `Item` read from the path.
 
         If an item cannot be read, return `None` instead and log an
@@ -1462,355 +1082,11 @@ class ImportTaskFactory:
                 # Silently ignore non-music files.
                 pass
             elif isinstance(exc.reason, mediafile.UnreadableFileError):
-                log.warning("unreadable file: {0}", displayable_path(path))
+                log.warning("unreadable file: {0}", util.displayable_path(path))
             else:
-                log.error("error reading {0}: {1}", displayable_path(path), exc)
-
-
-# Pipeline utilities
-
-
-def _freshen_items(items):
-    # Clear IDs from re-tagged items so they appear "fresh" when
-    # we add them back to the library.
-    for item in items:
-        item.id = None
-        item.album_id = None
-
-
-def _extend_pipeline(tasks, *stages):
-    # Return pipeline extension for stages with list of tasks
-    if isinstance(tasks, list):
-        task_iter = iter(tasks)
-    else:
-        task_iter = tasks
-
-    ipl = pipeline.Pipeline([task_iter] + list(stages))
-    return pipeline.multiple(ipl.pull())
-
-
-# Full-album pipeline stages.
-
-
-def read_tasks(session: ImportSession):
-    """A generator yielding all the albums (as ImportTask objects) found
-    in the user-specified list of paths. In the case of a singleton
-    import, yields single-item tasks instead.
-    """
-    skipped = 0
-
-    for toppath in session.paths:
-        # Check whether we need to resume the import.
-        session.ask_resume(toppath)
-
-        # Generate tasks.
-        task_factory = ImportTaskFactory(toppath, session)
-        yield from task_factory.tasks()
-        skipped += task_factory.skipped
-
-        if not task_factory.imported:
-            log.warning("No files imported from {0}", displayable_path(toppath))
-
-    # Show skipped directories (due to incremental/resume).
-    if skipped:
-        log.info("Skipped {0} paths.", skipped)
-
-
-def query_tasks(session: ImportSession):
-    """A generator that works as a drop-in-replacement for read_tasks.
-    Instead of finding files from the filesystem, a query is used to
-    match items from the library.
-    """
-    if session.config["singletons"]:
-        # Search for items.
-        for item in session.lib.items(session.query):
-            task = SingletonImportTask(None, item)
-            for task in task.handle_created(session):
-                yield task
-
-    else:
-        # Search for albums.
-        for album in session.lib.albums(session.query):
-            log.debug(
-                "yielding album {0}: {1} - {2}",
-                album.id,
-                album.albumartist,
-                album.album,
-            )
-            items = list(album.items())
-            _freshen_items(items)
-
-            task = ImportTask(None, [album.item_dir()], items)
-            for task in task.handle_created(session):
-                yield task
-
-
-@pipeline.mutator_stage
-def lookup_candidates(session: ImportSession, task: ImportTask):
-    """A coroutine for performing the initial MusicBrainz lookup for an
-    album. It accepts lists of Items and yields
-    (items, cur_artist, cur_album, candidates, rec) tuples. If no match
-    is found, all of the yielded parameters (except items) are None.
-    """
-    if task.skip:
-        # FIXME This gets duplicated a lot. We need a better
-        # abstraction.
-        return
-
-    plugins.send("import_task_start", session=session, task=task)
-    log.debug("Looking up: {0}", displayable_path(task.paths))
-
-    # Restrict the initial lookup to IDs specified by the user via the -m
-    # option. Currently all the IDs are passed onto the tasks directly.
-    task.search_ids = session.config["search_ids"].as_str_seq()
-
-    task.lookup_candidates()
-
-
-@pipeline.stage
-def user_query(session: ImportSession, task: ImportTask):
-    """A coroutine for interfacing with the user about the tagging
-    process.
-
-    The coroutine accepts an ImportTask objects. It uses the
-    session's `choose_match` method to determine the `action` for
-    this task. Depending on the action additional stages are executed
-    and the processed task is yielded.
-
-    It emits the ``import_task_choice`` event for plugins. Plugins have
-    access to the choice via the ``task.choice_flag`` property and may
-    choose to change it.
-    """
-    if task.skip:
-        return task
-
-    if session.already_merged(task.paths):
-        return pipeline.BUBBLE
-
-    # Ask the user for a choice.
-    task.choose_match(session)
-    plugins.send("import_task_choice", session=session, task=task)
-
-    # As-tracks: transition to singleton workflow.
-    if task.choice_flag is action.TRACKS:
-        # Set up a little pipeline for dealing with the singletons.
-        def emitter(task):
-            for item in task.items:
-                task = SingletonImportTask(task.toppath, item)
-                yield from task.handle_created(session)
-            yield SentinelImportTask(task.toppath, task.paths)
-
-        return _extend_pipeline(
-            emitter(task), lookup_candidates(session), user_query(session)
-        )
-
-    # As albums: group items by albums and create task for each album
-    if task.choice_flag is action.ALBUMS:
-        return _extend_pipeline(
-            [task],
-            group_albums(session),
-            lookup_candidates(session),
-            user_query(session),
-        )
-
-    resolve_duplicates(session, task)
-
-    if task.should_merge_duplicates:
-        # Create a new task for tagging the current items
-        # and duplicates together
-        duplicate_items = task.duplicate_items(session.lib)
-
-        # Duplicates would be reimported so make them look "fresh"
-        _freshen_items(duplicate_items)
-        duplicate_paths = [item.path for item in duplicate_items]
-
-        # Record merged paths in the session so they are not reimported
-        session.mark_merged(duplicate_paths)
-
-        merged_task = ImportTask(
-            None, task.paths + duplicate_paths, task.items + duplicate_items
-        )
-
-        return _extend_pipeline(
-            [merged_task], lookup_candidates(session), user_query(session)
-        )
-
-    apply_choice(session, task)
-    return task
-
-
-def resolve_duplicates(session: ImportSession, task: ImportTask):
-    """Check if a task conflicts with items or albums already imported
-    and ask the session to resolve this.
-    """
-    if task.choice_flag in (action.ASIS, action.APPLY, action.RETAG):
-        found_duplicates = task.find_duplicates(session.lib)
-        if found_duplicates:
-            log.debug(
-                "found duplicates: {}".format([o.id for o in found_duplicates])
-            )
-
-            # Get the default action to follow from config.
-            duplicate_action = config["import"]["duplicate_action"].as_choice(
-                {
-                    "skip": "s",
-                    "keep": "k",
-                    "remove": "r",
-                    "merge": "m",
-                    "ask": "a",
-                }
-            )
-            log.debug("default action for duplicates: {0}", duplicate_action)
-
-            if duplicate_action == "s":
-                # Skip new.
-                task.set_choice(action.SKIP)
-            elif duplicate_action == "k":
-                # Keep both. Do nothing; leave the choice intact.
-                pass
-            elif duplicate_action == "r":
-                # Remove old.
-                task.should_remove_duplicates = True
-            elif duplicate_action == "m":
-                # Merge duplicates together
-                task.should_merge_duplicates = True
-            else:
-                # No default action set; ask the session.
-                session.resolve_duplicate(task, found_duplicates)
-
-            session.log_choice(task, True)
-
-
-@pipeline.mutator_stage
-def import_asis(session: ImportSession, task: ImportTask):
-    """Select the `action.ASIS` choice for all tasks.
-
-    This stage replaces the initial_lookup and user_query stages
-    when the importer is run without autotagging.
-    """
-    if task.skip:
-        return
-
-    log.info("{}", displayable_path(task.paths))
-    task.set_choice(action.ASIS)
-    apply_choice(session, task)
-
-
-def apply_choice(session: ImportSession, task: ImportTask):
-    """Apply the task's choice to the Album or Item it contains and add
-    it to the library.
-    """
-    if task.skip:
-        return
-
-    # Change metadata.
-    if task.apply:
-        task.apply_metadata()
-        plugins.send("import_task_apply", session=session, task=task)
-
-    task.add(session.lib)
-
-    # If ``set_fields`` is set, set those fields to the
-    # configured values.
-    # NOTE: This cannot be done before the ``task.add()`` call above,
-    # because then the ``ImportTask`` won't have an `album` for which
-    # it can set the fields.
-    if config["import"]["set_fields"]:
-        task.set_fields(session.lib)
-
-
-@pipeline.mutator_stage
-def plugin_stage(
-    session: ImportSession,
-    func: Callable[[ImportSession, ImportTask], None],
-    task: ImportTask,
-):
-    """A coroutine (pipeline stage) that calls the given function with
-    each non-skipped import task. These stages occur between applying
-    metadata changes and moving/copying/writing files.
-    """
-    if task.skip:
-        return
-
-    func(session, task)
-
-    # Stage may modify DB, so re-load cached item data.
-    # FIXME Importer plugins should not modify the database but instead
-    # the albums and items attached to tasks.
-    task.reload()
-
-
-@pipeline.stage
-def manipulate_files(session: ImportSession, task: ImportTask):
-    """A coroutine (pipeline stage) that performs necessary file
-    manipulations *after* items have been added to the library and
-    finalizes each task.
-    """
-    if not task.skip:
-        if task.should_remove_duplicates:
-            task.remove_duplicates(session.lib)
-
-        if session.config["move"]:
-            operation = MoveOperation.MOVE
-        elif session.config["copy"]:
-            operation = MoveOperation.COPY
-        elif session.config["link"]:
-            operation = MoveOperation.LINK
-        elif session.config["hardlink"]:
-            operation = MoveOperation.HARDLINK
-        elif session.config["reflink"] == "auto":
-            operation = MoveOperation.REFLINK_AUTO
-        elif session.config["reflink"]:
-            operation = MoveOperation.REFLINK
-        else:
-            operation = None
-
-        task.manipulate_files(
-            session=session,
-            operation=operation,
-            write=session.config["write"],
-        )
-
-    # Progress, cleanup, and event.
-    task.finalize(session)
-
-
-@pipeline.stage
-def log_files(session: ImportSession, task: ImportTask):
-    """A coroutine (pipeline stage) to log each file to be imported."""
-    if isinstance(task, SingletonImportTask):
-        log.info("Singleton: {0}", displayable_path(task.item["path"]))
-    elif task.items:
-        log.info("Album: {0}", displayable_path(task.paths[0]))
-        for item in task.items:
-            log.info("  {0}", displayable_path(item["path"]))
-
-
-def group_albums(session: ImportSession):
-    """A pipeline stage that groups the items of each task into albums
-    using their metadata.
-
-    Groups are identified using their artist and album fields. The
-    pipeline stage emits new album tasks for each discovered group.
-    """
-
-    def group(item):
-        return (item.albumartist or item.artist, item.album)
-
-    task = None
-    while True:
-        task = yield task
-        if task.skip:
-            continue
-        tasks = []
-        sorted_items: list[library.Item] = sorted(task.items, key=group)
-        for _, items in itertools.groupby(sorted_items, group):
-            l_items = list(items)
-            task = ImportTask(task.toppath, [i.path for i in l_items], l_items)
-            tasks += task.handle_created(session)
-        tasks.append(SentinelImportTask(task.toppath, task.paths))
-
-        task = pipeline.multiple(tasks)
+                log.error(
+                    "error reading {0}: {1}", util.displayable_path(path), exc
+                )
 
 
 MULTIDISC_MARKERS = (rb"dis[ck]", rb"cd")
@@ -1821,11 +1097,11 @@ def is_subdir_of_any_in_list(path, dirs):
     """Returns True if path os a subdirectory of any directory in dirs
     (a list). In other case, returns False.
     """
-    ancestors = ancestry(path)
+    ancestors = util.ancestry(path)
     return any(d in ancestors for d in dirs)
 
 
-def albums_in_dir(path: PathBytes):
+def albums_in_dir(path: util.PathBytes):
     """Recursively searches the given directory and returns an iterable
     of (paths, items) where paths is a list of directories and items is
     a list of Items that is probably an album. Specifically, any folder
@@ -1835,7 +1111,7 @@ def albums_in_dir(path: PathBytes):
     ignore: list[str] = config["ignore"].as_str_seq()
     ignore_hidden: bool = config["ignore_hidden"].get(bool)
 
-    for root, dirs, files in sorted_walk(
+    for root, dirs, files in util.sorted_walk(
         path, ignore=ignore, ignore_hidden=ignore_hidden, logger=log
     ):
         items = [os.path.join(root, f) for f in files]
